+++ src/aimbat/_cli/__init__.py
@@ -1,13 +1,14 @@
 from .align import app as align
 from .data import app as data
 from .event import app as event
-from .pick import app as pick
 from .plot import app as plot
 from .project import app as project
 from .seismogram import app as seismogram
 from .shell import app as shell
 from .snapshot import app as snapshot
 from .station import app as station
+from .tool import app as tool
+from .tui import app as tui
 from .utils import app as utils
 
 __all__ = [
@@ -15,11 +16,12 @@ __all__ = [
     "align",
     "data",
     "event",
-    "pick",
+    "tool",
     "plot",
     "project",
     "seismogram",
     "shell",
     "snapshot",
     "station",
+    "tui",
 ]